refac: repository interface
