+++ src/main/java/roomescape/common/exception/BeforeOpenException.java
