+++ src/main/java/roomescape/common/exception/GlobalExceptionHandler.java
@@ -60,4 +60,3 @@ public class GlobalExceptionHandler {
                 .body(ErrorResponseBody.from(e.getMessage()));
     }
 }
-
