--- src/main/java/roomescape/reservation/repository/ReservationUpdatingDAO.java
@@ -1,41 +1,0 @@
-package roomescape.reservation.repository;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-import roomescape.reservation.model.Reservation;
-
-import java.sql.PreparedStatement;
-
-@Repository
-public class ReservationUpdatingDAO {
-
-    private JdbcTemplate jdbcTemplate;
-
-    public ReservationUpdatingDAO(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public Long insertWithKeyHolder(Reservation reservation) {
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(
-                    "insert into reservation (name, date, time_id) values (?, ?, ?)",
-                    new String[]{"id"}
-            );
-            ps.setString(1, reservation.getName());
-            ps.setString(2, reservation.getDate().toString());
-            ps.setLong(3, reservation.getTime().getId());
-            return ps;
-        }, keyHolder);
-
-        return keyHolder.getKey().longValue();
-    }
-
-    public int delete(Reservation reservation) {
-        String sql = "delete from reservation where id = ?";
-        Long id = reservation.getId();
-        return jdbcTemplate.update(sql, id);
-    }
-}
