+++ src/main/java/roomescape/time/service/TimeService.java
@@ -7,45 +7,42 @@ import roomescape.time.dto.TimeCreationRequest;
 import roomescape.time.dto.TimeResponse;
 import roomescape.time.model.Time;
 import roomescape.time.model.TimeManager;
-import roomescape.time.repository.TimeQueryingDAO;
-import roomescape.time.repository.TimeUpdatingDAO;
+import roomescape.time.repository.TimeRepository;
 
 import java.util.List;
 
 @Service
 public class TimeService {
 
-    private final TimeQueryingDAO timeQueryingDAO;
-    private final TimeUpdatingDAO timeUpdatingDAO;
+    private final TimeRepository timeRepository;
 
     @Autowired
-    public TimeService(TimeQueryingDAO timeQueryingDAO, TimeUpdatingDAO timeUpdatingDAO) {
-        this.timeQueryingDAO = timeQueryingDAO;
-        this.timeUpdatingDAO = timeUpdatingDAO;
+    public TimeService(TimeRepository timeRepository) {
+        this.timeRepository = timeRepository;
     }
 
     public List<TimeResponse> getTimeList() {
-        List<Time> timeList = timeQueryingDAO.findAllTime();
+        List<Time> timeList = timeRepository.findAllTime();
 
         return timeList.stream().map(TimeResponse::from).toList();
     }
 
     public TimeResponse addTime(TimeCreationRequest request) {
-        List<Time> timeList = timeQueryingDAO.findAllTime();
+        List<Time> timeList = timeRepository.findAllTime();
         TimeManager timeManager = new TimeManager(timeList);
 
         Time newTime = TimeCreationRequest.toEntityFrom(request);
 
         timeManager.validateTime(newTime);
-        Long id = timeUpdatingDAO.insertWithKeyHolder(newTime);
+        Long id = timeRepository.insertWithKeyHolder(newTime);
         newTime.setId(id);
 
         return TimeResponse.from(newTime);
     }
 
     public void deleteTime(Long id) {
-        Time time = timeQueryingDAO.findTimeById(id)
+        Time time = timeRepository.findTimeById(id)
                 .orElseThrow(() -> new NoSuchElementToDeleteException("There is no reservation with id " + id));
-        timeUpdatingDAO.delete(time);
+        timeRepository.delete(time);
     }
 }
